Android Based Encryption SMS System
The project focuses on the basic sms transfer from one sender to another safely. The "safe" word describes that any intruder should not be able to retrieve the message from the known attacks in cryptography, hence making it trustable source of communication. The application consists of a instant and secure login. The application, then asks the user whether he/she wants to craft a message or see his inbox to figure out what's new? The crafting of message includes the feature of accessing the contacts. The user can instantly see his message and the corresponding cipher text. There is a tutorial accompanied to understand how the encryption and decryption is being done. The receiver finally will receive the sms which this application will decrypt and store in its inbox. The app also supports group messaging, wherein a common message can be sent to a group of receivers.